Windows C4267 errors: Fix size_t conversion warnings in windows builds
examples/client.c:
examples/coap-rd.c:
examples/coap-server.c:
include/coap2/block.h:
include/coap2/coap_hashkey.h:
src/block.c:
src/coap_debug.c:
src/coap_hashkey.c:
src/coap_io.c:
src/coap_openssl.c:
src/net.c:

Fix C4267 windows compilation warnings when using Visual Studio where
size_t variable parameters are being converted  in both size and sign.

There still are some C4116 warnings to fix.

+++ include/coap2/coap_hashkey.h
@@ -31,7 +31,7 @@ typedef unsigned char coap_key_t[4];
  * @param len The length of @p s.
  * @param h   The result buffer to store the calculated hash key.
  */
-void coap_hash_impl(const unsigned char *s, unsigned int len, coap_key_t h);
+void coap_hash_impl(const unsigned char *s, size_t len, coap_key_t h);
 
 #define coap_hash(String,Length,Result) \
   coap_hash_impl((String),(Length),(Result))
