+++ src/coap_hashkey.c
@@ -9,7 +9,7 @@
 #include "coap_internal.h"
 
 void
-coap_hash_impl(const unsigned char *s, unsigned int len, coap_key_t h) {
+coap_hash_impl(const unsigned char *s, size_t len, coap_key_t h) {
   size_t j;
 
   while (len--) {
